services/api-gateway/requirements.txt
django==5.1
djangorestframework==3.15.2
grpcio==1.76.0
protobuf==6.33.4
